+++ module/applications/compendium/compendium-browser.mjs
@@ -1,5 +1,5 @@
 import { Progress } from "./progress.mjs";
-import { fontAwesomeIcon, htmlQueryAll, objectHasKey, isObject, getSelectedOrOwnActors } from "../../utils.mjs";
+import { fontAwesomeIcon, htmlQuery, htmlQueryAll, objectHasKey, isObject, getSelectedOrOwnActors } from "../../utils.mjs";
 import * as browserTabs from "./tabs/_module.mjs";
 import Tagify from "@yaireo/tagify";
 
@@ -118,16 +118,16 @@ export default class CompendiumBrowser extends Application {
           navSelector: "nav",
           contentSelector: "section.content",
           initial: "landing-page"
+        },
+        {
+          navSelector: "nav[data-group=settings]",
+          contentSelector: ".settings-container",
+          initial: "packs",
         }
       ],
-      scrollY: [".control-area", ".item-list"]
+      scrollY: [".control-area", ".item-list", ".settings-container"]
     });
   }
-
-  // /** @inheritdoc */
-  // async render(force, options) {
-  //   return super.render(force, options);
-  // }
 
   /**
    * Reset initial filtering
@@ -216,54 +216,28 @@ export default class CompendiumBrowser extends Application {
       types.delete(undefined);
       if (types.size === 0) continue;
 
-      if (types.has("npc")) {
-        const load = this.settings.bestiary?.[pack.collection]?.load ?? true;
-        settings.bestiary[pack.collection] = {
-          load,
-          name: pack.metadata.label
-        };
-      }
-      if (types.has("starship")) {
-        const load = this.settings.shipyard?.[pack.collection]?.load ?? true;
-        settings.shipyard[pack.collection] = {
-          load,
-          name: pack.metadata.label
-        };
-      }
+      const type = (() => {
+        if ( types.has("npc") ) return "bestiary";
 
-      let t = CONFIG.SW5E.itemTypes.inventory.find(type => types.has(type));
-      if (t !== undefined) {
-        const load = this.settings.equipment?.[pack.collection]?.load ?? !!loadDefault[pack.collection];
-        settings.equipment[pack.collection] = {
-          load,
-          name: pack.metadata.label
-        };
-        continue;
-      }
+        if ( types.has("starship") ) return "shipyard";
 
-      t = CONFIG.SW5E.itemTypes.class.find(type => types.has(type));
-      if (t !== undefined) {
-        const load = this.settings.class?.[pack.collection]?.load ?? !!loadDefault[pack.collection];
-        settings.class[pack.collection] = {
-          load,
-          name: pack.metadata.label
-        };
-      }
+        if ( CONFIG.SW5E.itemTypes.inventory.some((type) => types.has(type)) ) return "equipment";
 
-      t = CONFIG.SW5E.itemTypes.other.find(type => types.has(type));
-      if (t !== undefined) {
-        const load = this.settings.other?.[pack.collection]?.load ?? !!loadDefault[pack.collection];
-        settings.other[pack.collection] = {
-          load,
-          name: pack.metadata.label
-        };
-      }
+        if ( CONFIG.SW5E.itemTypes.class.some((type) => types.has(type)) ) return "class";
 
-      if (types.has("feat")) {
-        const load = this.settings.feat?.[pack.collection]?.load ?? !!loadDefault[pack.collection];
-        settings.feat[pack.collection] = {
+        if ( types.has("feat") ) return "feat";
+
+        if ( CONFIG.SW5E.itemTypes.other.some((type) => types.has(type)) ) return "other";
+
+        return null;
+      })();
+
+      if (type) {
+        const load = this.settings[type]?.[pack.collection]?.load ?? loadDefault[type] ?? !!loadDefault[pack.collection];
+        settings[type][pack.collection] = {
           load,
-          name: pack.metadata.label
+          name: pack.metadata.label,
+          package: pack.metadata.packageName,
         };
       }
     }
@@ -290,7 +264,7 @@ export default class CompendiumBrowser extends Application {
     return this.loadTab(tabName);
   }
 
-  async openPowerTab(entry, maxLevel = 10) {
+  async openPowerTab(entry, maxLevel = 9) {
     const powerTab = this.tabs.power;
     const filter = await powerTab.getFilterData();
     const { category, level, traditions } = filter.checkboxes;
@@ -324,6 +298,7 @@ export default class CompendiumBrowser extends Application {
     this.activeTab = tabName;
     // Settings tab
     if (tabName === "settings") {
+      await this.packLoader.updateSources(this.loadedPacksAll());
       await this.render(true);
       return;
     }
@@ -347,10 +322,18 @@ export default class CompendiumBrowser extends Application {
     });
   }
 
+  loadedPacksAll() {
+    const loadedPacks = new Set();
+    for (const tabName of this.dataTabsList) {
+      this.loadedPacks(tabName).forEach((item) => loadedPacks.add(item));
+    }
+    return Array.from(loadedPacks).sort();
+  }
+
   /** @inheritdoc */
-  activateListeners(html) {
-    super.activateListeners(html);
-    html = html[0];
+  activateListeners(input_html) {
+    super.activateListeners(input_html);
+    const html = input_html[0];
     const activeTabName = this.activeTab;
 
     // Set the navigation tab. This is only needed when the browser is openend
@@ -359,25 +342,87 @@ export default class CompendiumBrowser extends Application {
 
     // Settings Tab
     if (activeTabName === "settings") {
-      const form = html.querySelector(".compendium-browser-settings form");
-      if (form) {
-        form.querySelector("button.save-settings")?.addEventListener("click", async () => {
-          const formData = new FormData(form);
-          for (const [t, packs] of Object.entries(this.settings)) {
-            for (const [key, pack] of Object.entries(packs)) {
-              pack.load = formData.has(`${t}-${key}`);
-            }
+      const settings = htmlQuery(html, ".compendium-browser-settings");
+      const form = settings?.querySelector("form");
+      if (!form) return;
+
+      htmlQuery(settings, "button[data-action=save-settings]")?.addEventListener("click", async () => {
+        const formData = new FormData(form);
+        for (const [t, packs] of Object.entries(this.settings)) {
+          for (const [key, pack] of Object.entries(packs)) {
+            pack.load = formData.has(`${t}-${key}`);
           }
-          await game.settings.set("sw5e", "compendiumBrowserPacks", this.settings);
-          for (const tab of Object.values(this.tabs)) {
-            if (tab.isInitialized) {
-              await tab.init();
-              tab.scrollLimit = 100;
-            }
+        }
+        await game.settings.set("sw5e", "compendiumBrowserPacks", this.settings);
+
+        for (const [key, source] of Object.entries(this.packLoader.sourcesSettings.sources)) {
+          if (!source || isBlank(source.name)) {
+            delete this.packLoader.sourcesSettings.sources[key]; // just to make sure we clean up
+            continue;
           }
-          this.render(true);
+          source.load = formData.has(`source-${key}`);
+        }
+
+        this.packLoader.sourcesSettings.showEmptySources = formData.has("show-empty-sources");
+        this.packLoader.sourcesSettings.showUnknownSources = formData.has("show-unknown-sources");
+        this.packLoader.sourcesSettings.ignoreAsGM = formData.has("ignore-as-gm");
+        await game.settings.set("sw5e", "compendiumBrowserSources", this.packLoader.sourcesSettings);
+
+        await this.#resetInitializedTabs();
+        this.render(true);
+        ui.notifications.info("SW5E.BrowserSettingsSaved", { localize: true });
+      });
+
+      const sourceSearch = htmlQuery(form, "input[data-element=setting-sources-search]");
+      const sourceToggle = htmlQuery(form, "input[data-action=setting-sources-toggle-visible]");
+      const sourceSettings = htmlQueryAll(form, "label[data-element=setting-source]");
+
+      sourceSearch?.addEventListener("input", () => {
+        const value = sourceSearch.value?.trim().toLocaleLowerCase(game.i18n.lang);
+
+        for (const element of sourceSettings) {
+          const name = element.dataset.name?.toLocaleLowerCase(game.i18n.lang);
+          const shouldBeHidden = !isBlank(value) && !isBlank(name) && !name.includes(value);
+
+          element.classList.toggle("hidden", shouldBeHidden);
+        }
+
+        if (sourceToggle) {
+          sourceToggle.checked = false;
+        }
+      });
+
+      sourceToggle?.addEventListener("click", () => {
+        for (const element of sourceSettings) {
+          const checkbox = htmlQuery(element, "input[type=checkbox]");
+          if (!element.classList.contains("hidden") && checkbox) {
+            checkbox.checked = sourceToggle.checked;
+          }
+        }
+      });
+
+      const deleteButton = htmlQuery(form, "button[data-action=settings-sources-delete]");
+      deleteButton?.addEventListener("click", async () => {
+        const localize = localizer("SW5E.SETTINGS.CompendiumBrowserSources");
+        const confirm = await Dialog.confirm({
+          title: localize("DeleteAllTitle"),
+          content: `
+            <p>
+              ${localize("DeleteAllQuestion")}
+            </p>
+            <p>
+              ${localize("DeleteAllInfo")}
+            </p>
+            `,
         });
-      }
+
+        if (confirm) {
+          await this.packLoader.hardReset(this.loadedPacksAll());
+          await game.settings.set("sw5e", "compendiumBrowserSources", this.packLoader.sourcesSettings);
+          await this.#resetInitializedTabs();
+          this.render(true);
+        }
+      });
       return;
     }
 
@@ -434,6 +479,17 @@ export default class CompendiumBrowser extends Application {
     controlArea.querySelector("button.clear-filters")?.addEventListener("click", () => {
       this.resetFilters();
       this.clearScrollLimit(true);
+    });
+
+    // Create Roll Table button
+    htmlQuery(html, "[data-action=create-roll-table]")?.addEventListener("click", () =>
+      currentTab.createRollTable()
+    );
+
+    // Add to Roll Table button
+    htmlQuery(html, "[data-action=add-to-roll-table]")?.addEventListener("click", async () => {
+      if (!game.tables.contents.length) return;
+      currentTab.addToRollTable();
     });
 
     // Filters
@@ -640,8 +696,8 @@ export default class CompendiumBrowser extends Application {
             data.values.min = min;
             data.values.max = max;
 
-            const minLabel = html.find(`label.${name}-min-label`);
-            const maxLabel = html.find(`label.${name}-max-label`);
+            const minLabel = input_html.find(`label.${name}-min-label`);
+            const maxLabel = input_html.find(`label.${name}-max-label`);
             minLabel.text(min);
             maxLabel.text(max);
 
@@ -674,6 +730,15 @@ export default class CompendiumBrowser extends Application {
 
     // Initial result list render
     this.renderResultList({ list });
+  }
+
+  async #resetInitializedTabs() {
+    for (const tab of Object.values(this.tabs)) {
+      if (tab.isInitialized) {
+        await tab.init();
+        tab.scrollLimit = 100;
+      }
+    }
   }
 
   /**
@@ -838,6 +903,9 @@ export default class CompendiumBrowser extends Application {
         uuid: item.dataset.entryUuid
       })
     );
+    // awful hack (dataTransfer.types will include "from-browser")
+    event.dataTransfer.setData("from-browser", "true");
+
     item.addEventListener(
       "dragend",
       () => {
@@ -853,25 +921,25 @@ export default class CompendiumBrowser extends Application {
 
   _onDragOver(event) {
     super._onDragOver(event);
-    this.element.css({ pointerEvents: "none" });
+    if (event.dataTransfer.types.includes("from-browser")) {
+      this.element.css({ pointerEvents: "none" });
+    }
   }
 
   getData() {
     const activeTab = this.activeTab;
-    // Settings
-    if (activeTab === "settings") return {
-      user: game.user,
-      settings: this.settings
+    const tab = objectHasKey(this.tabs, activeTab) ? this.tabs[activeTab] : null;
+
+    const settings = {
+      settings: this.settings,
+      sources: this.packLoader.sourcesSettings,
     };
-    // Active tab
-    const tab = this.tabs[activeTab];
-    if (tab) return {
+
+    return {
       user: game.user,
-      [activeTab]: { filterData: tab.filterData },
-      scrollLimit: tab.scrollLimit
+      [activeTab]: activeTab === "settings" ? settings : { filterData: tab?.filterData },
+      scrollLimit: tab?.scrollLimit,
     };
-    // No active tab
-    return { user: game.user };
   }
 
   resetFilters() {